Struct aws_sdk_route53::client::fluent_builders::CreateKeySigningKey

pub struct CreateKeySigningKey<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateKeySigningKey<C>

pub async fn send(
self
) -> Result<CreateKeySigningKeyOutput, SdkError<CreateKeySigningKeyError>> where
C: SmithyConnector,

pub fn caller_reference(self, input: impl Into<String>) -> Self

A unique string that identifies the request.

pub fn set_caller_reference(self, input: Option<String>) -> Self

pub fn hosted_zone_id(self, input: impl Into<String>) -> Self

The unique string (ID) used to identify a hosted zone.

pub fn set_hosted_zone_id(self, input: Option<String>) -> Self

pub fn key_management_service_arn(self, input: impl Into<String>) -> Self

The Amazon resource name (ARN) for a customer managed customer master key (CMK) in AWS Key Management Service (AWS KMS). The KeyManagementServiceArn must be unique for each key-signing key (KSK) in a single hosted zone. To see an example of KeyManagementServiceArn that grants the correct permissions for DNSSEC, scroll down to Example.

You must configure the customer managed CMK as follows:

Status

Enabled

Key spec

ECC_NIST_P256

Key usage

Sign and verify

Key policy

The key policy must give permission for the following actions:

DescribeKey

GetPublicKey

Sign

The key policy must also include the Amazon Route 53 service in the principal for your account. Specify the following:

"Service": "dnssec.route53.aws.amazonaws.com"

For more information about working with a customer managed CMK in AWS KMS, see AWS Key Management Service concepts.

pub fn set_key_management_service_arn(self, input: Option<String>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

A string used to identify a key-signing key (KSK). Name can include numbers, letters, and underscores (_). Name must be unique for each key-signing key in the same hosted zone.

pub fn set_name(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

A string specifying the initial status of the key-signing key (KSK). You can set the value to ACTIVE or INACTIVE.

pub fn set_status(self, input: Option<String>) -> Self