Struct aws_sdk_route53::client::fluent_builders::ListResourceRecordSets

pub struct ListResourceRecordSets<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> ListResourceRecordSets<C>

pub async fn send(
self
) -> Result<ListResourceRecordSetsOutput, SdkError<ListResourceRecordSetsError>> where
C: SmithyConnector,

pub fn hosted_zone_id(self, input: impl Into<String>) -> Self

The ID of the hosted zone that contains the resource record sets that you want to list.

pub fn set_hosted_zone_id(self, input: Option<String>) -> Self

pub fn start_record_name(self, input: impl Into<String>) -> Self

The first name in the lexicographic ordering of resource record sets that you want to list. If the specified record name doesn't exist, the results begin with the first resource record set that has a name greater than the value of name.

pub fn set_start_record_name(self, input: Option<String>) -> Self

pub fn start_record_type(self, input: RrType) -> Self

The type of resource record set to begin the record listing from.

Valid values for basic resource record sets: A | AAAA | CAA | CNAME | MX | NAPTR | NS | PTR | SOA | SPF | SRV | TXT

Values for weighted, latency, geolocation, and failover resource record sets: A | AAAA | CAA | CNAME | MX | NAPTR | PTR | SPF | SRV | TXT

Values for alias resource record sets:

API Gateway custom regional API or edge-optimized API: A

CloudFront distribution: A or AAAA

Elastic Beanstalk environment that has a regionalized subdomain: A

Elastic Load Balancing load balancer: A | AAAA

S3 bucket: A

VPC interface VPC endpoint: A

Another resource record set in this hosted zone: The type of the resource record set that the alias references.

Constraint: Specifying type without specifying name returns an InvalidInput error.

pub fn set_start_record_type(self, input: Option<RrType>) -> Self

pub fn start_record_identifier(self, input: impl Into<String>) -> Self

Resource record sets that have a routing policy other than simple: If results were truncated for a given DNS name and type, specify the value of NextRecordIdentifier from the previous response to get the next resource record set that has the current DNS name and type.

pub fn set_start_record_identifier(self, input: Option<String>) -> Self

pub fn max_items(self, input: i32) -> Self

(Optional) The maximum number of resource records sets to include in the response body for this request. If the response includes more than maxitems resource record sets, the value of the IsTruncated element in the response is true, and the values of the NextRecordName and NextRecordType elements in the response identify the first resource record set in the next group of maxitems resource record sets.

pub fn set_max_items(self, input: Option<i32>) -> Self