Struct aws_sdk_route53::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSDnsV20130401.

This client allows ergonomic access to a AWSDnsV20130401-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn activate_key_signing_key(&self) -> ActivateKeySigningKey<C, M, R>

pub fn associate_vpc_with_hosted_zone(
&self
) -> AssociateVPCWithHostedZone<C, M, R>

pub fn change_resource_record_sets(&self) -> ChangeResourceRecordSets<C, M, R>

pub fn change_tags_for_resource(&self) -> ChangeTagsForResource<C, M, R>

pub fn create_health_check(&self) -> CreateHealthCheck<C, M, R>

pub fn create_hosted_zone(&self) -> CreateHostedZone<C, M, R>

pub fn create_key_signing_key(&self) -> CreateKeySigningKey<C, M, R>

pub fn create_query_logging_config(&self) -> CreateQueryLoggingConfig<C, M, R>

pub fn create_reusable_delegation_set(
&self
) -> CreateReusableDelegationSet<C, M, R>

pub fn create_traffic_policy(&self) -> CreateTrafficPolicy<C, M, R>

pub fn create_traffic_policy_instance(
&self
) -> CreateTrafficPolicyInstance<C, M, R>

pub fn create_traffic_policy_version(
&self
) -> CreateTrafficPolicyVersion<C, M, R>

pub fn create_vpc_association_authorization(
&self
) -> CreateVPCAssociationAuthorization<C, M, R>

pub fn deactivate_key_signing_key(&self) -> DeactivateKeySigningKey<C, M, R>

pub fn delete_health_check(&self) -> DeleteHealthCheck<C, M, R>

pub fn delete_hosted_zone(&self) -> DeleteHostedZone<C, M, R>

pub fn delete_key_signing_key(&self) -> DeleteKeySigningKey<C, M, R>

pub fn delete_query_logging_config(&self) -> DeleteQueryLoggingConfig<C, M, R>

pub fn delete_reusable_delegation_set(
&self
) -> DeleteReusableDelegationSet<C, M, R>

pub fn delete_traffic_policy(&self) -> DeleteTrafficPolicy<C, M, R>

pub fn delete_traffic_policy_instance(
&self
) -> DeleteTrafficPolicyInstance<C, M, R>

pub fn delete_vpc_association_authorization(
&self
) -> DeleteVPCAssociationAuthorization<C, M, R>

pub fn disable_hosted_zone_dnssec(&self) -> DisableHostedZoneDNSSEC<C, M, R>

pub fn disassociate_vpc_from_hosted_zone(
&self
) -> DisassociateVPCFromHostedZone<C, M, R>

pub fn enable_hosted_zone_dnssec(&self) -> EnableHostedZoneDNSSEC<C, M, R>

pub fn get_account_limit(&self) -> GetAccountLimit<C, M, R>

pub fn get_change(&self) -> GetChange<C, M, R>

pub fn get_checker_ip_ranges(&self) -> GetCheckerIpRanges<C, M, R>

pub fn get_dnssec(&self) -> GetDNSSEC<C, M, R>

pub fn get_geo_location(&self) -> GetGeoLocation<C, M, R>

pub fn get_health_check(&self) -> GetHealthCheck<C, M, R>

pub fn get_health_check_count(&self) -> GetHealthCheckCount<C, M, R>

pub fn get_health_check_last_failure_reason(
&self
) -> GetHealthCheckLastFailureReason<C, M, R>

pub fn get_health_check_status(&self) -> GetHealthCheckStatus<C, M, R>

pub fn get_hosted_zone(&self) -> GetHostedZone<C, M, R>

pub fn get_hosted_zone_count(&self) -> GetHostedZoneCount<C, M, R>

pub fn get_hosted_zone_limit(&self) -> GetHostedZoneLimit<C, M, R>

pub fn get_query_logging_config(&self) -> GetQueryLoggingConfig<C, M, R>

pub fn get_reusable_delegation_set(&self) -> GetReusableDelegationSet<C, M, R>

pub fn get_reusable_delegation_set_limit(
&self
) -> GetReusableDelegationSetLimit<C, M, R>

pub fn get_traffic_policy(&self) -> GetTrafficPolicy<C, M, R>

pub fn get_traffic_policy_instance(&self) -> GetTrafficPolicyInstance<C, M, R>

pub fn get_traffic_policy_instance_count(
&self
) -> GetTrafficPolicyInstanceCount<C, M, R>

pub fn list_geo_locations(&self) -> ListGeoLocations<C, M, R>

pub fn list_health_checks(&self) -> ListHealthChecks<C, M, R>

pub fn list_hosted_zones(&self) -> ListHostedZones<C, M, R>

pub fn list_hosted_zones_by_name(&self) -> ListHostedZonesByName<C, M, R>

pub fn list_hosted_zones_by_vpc(&self) -> ListHostedZonesByVPC<C, M, R>

pub fn list_query_logging_configs(&self) -> ListQueryLoggingConfigs<C, M, R>

pub fn list_resource_record_sets(&self) -> ListResourceRecordSets<C, M, R>

pub fn list_reusable_delegation_sets(
&self
) -> ListReusableDelegationSets<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_tags_for_resources(&self) -> ListTagsForResources<C, M, R>

pub fn list_traffic_policies(&self) -> ListTrafficPolicies<C, M, R>

pub fn list_traffic_policy_instances(
&self
) -> ListTrafficPolicyInstances<C, M, R>

pub fn list_traffic_policy_instances_by_hosted_zone(
&self
) -> ListTrafficPolicyInstancesByHostedZone<C, M, R>

pub fn list_traffic_policy_instances_by_policy(
&self
) -> ListTrafficPolicyInstancesByPolicy<C, M, R>

pub fn list_traffic_policy_versions(&self) -> ListTrafficPolicyVersions<C, M, R>

pub fn list_vpc_association_authorizations(
&self
) -> ListVPCAssociationAuthorizations<C, M, R>

pub fn test_dns_answer(&self) -> TestDNSAnswer<C, M, R>

pub fn update_health_check(&self) -> UpdateHealthCheck<C, M, R>

pub fn update_hosted_zone_comment(&self) -> UpdateHostedZoneComment<C, M, R>

pub fn update_traffic_policy_comment(
&self
) -> UpdateTrafficPolicyComment<C, M, R>

pub fn update_traffic_policy_instance(
&self
) -> UpdateTrafficPolicyInstance<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self