Enum aws_sdk_route53::Error

#[non_exhaustive]
pub enum Error {

    ConcurrentModification(ConcurrentModification),
    ConflictingDomainExists(ConflictingDomainExists),
    ConflictingTypes(ConflictingTypes),
    DnssecNotFound(DnssecNotFound),
    DelegationSetAlreadyCreated(DelegationSetAlreadyCreated),
    DelegationSetAlreadyReusable(DelegationSetAlreadyReusable),
    DelegationSetInUse(DelegationSetInUse),
    DelegationSetNotAvailable(DelegationSetNotAvailable),
    DelegationSetNotReusable(DelegationSetNotReusable),
    HealthCheckAlreadyExists(HealthCheckAlreadyExists),
    HealthCheckInUse(HealthCheckInUse),
    HealthCheckVersionMismatch(HealthCheckVersionMismatch),
    HostedZoneAlreadyExists(HostedZoneAlreadyExists),
    HostedZoneNotEmpty(HostedZoneNotEmpty),
    HostedZoneNotFound(HostedZoneNotFound),
    HostedZoneNotPrivate(HostedZoneNotPrivate),
    HostedZonePartiallyDelegated(HostedZonePartiallyDelegated),
    IncompatibleVersion(IncompatibleVersion),
    InsufficientCloudWatchLogsResourcePolicy(InsufficientCloudWatchLogsResourcePolicy),
    InvalidArgument(InvalidArgument),
    InvalidChangeBatch(InvalidChangeBatch),
    InvalidDomainName(InvalidDomainName),
    InvalidInput(InvalidInput),
    InvalidKmsArn(InvalidKmsArn),
    InvalidKeySigningKeyName(InvalidKeySigningKeyName),
    InvalidKeySigningKeyStatus(InvalidKeySigningKeyStatus),
    InvalidPaginationToken(InvalidPaginationToken),
    InvalidSigningStatus(InvalidSigningStatus),
    InvalidTrafficPolicyDocument(InvalidTrafficPolicyDocument),
    InvalidVpcId(InvalidVpcId),
    KeySigningKeyAlreadyExists(KeySigningKeyAlreadyExists),
    KeySigningKeyInParentDsRecord(KeySigningKeyInParentDsRecord),
    KeySigningKeyInUse(KeySigningKeyInUse),
    KeySigningKeyWithActiveStatusNotFound(KeySigningKeyWithActiveStatusNotFound),
    LastVpcAssociation(LastVpcAssociation),
    LimitsExceeded(LimitsExceeded),
    NoSuchChange(NoSuchChange),
    NoSuchCloudWatchLogsLogGroup(NoSuchCloudWatchLogsLogGroup),
    NoSuchDelegationSet(NoSuchDelegationSet),
    NoSuchGeoLocation(NoSuchGeoLocation),
    NoSuchHealthCheck(NoSuchHealthCheck),
    NoSuchHostedZone(NoSuchHostedZone),
    NoSuchKeySigningKey(NoSuchKeySigningKey),
    NoSuchQueryLoggingConfig(NoSuchQueryLoggingConfig),
    NoSuchTrafficPolicy(NoSuchTrafficPolicy),
    NoSuchTrafficPolicyInstance(NoSuchTrafficPolicyInstance),
    NotAuthorizedException(NotAuthorizedException),
    PriorRequestNotComplete(PriorRequestNotComplete),
    PublicZoneVpcAssociation(PublicZoneVpcAssociation),
    QueryLoggingConfigAlreadyExists(QueryLoggingConfigAlreadyExists),
    ThrottlingException(ThrottlingException),
    TooManyHealthChecks(TooManyHealthChecks),
    TooManyHostedZones(TooManyHostedZones),
    TooManyKeySigningKeys(TooManyKeySigningKeys),
    TooManyTrafficPolicies(TooManyTrafficPolicies),
    TooManyTrafficPolicyInstances(TooManyTrafficPolicyInstances),
    TooManyTrafficPolicyVersionsForCurrentPolicy(TooManyTrafficPolicyVersionsForCurrentPolicy),
    TooManyVpcAssociationAuthorizations(TooManyVpcAssociationAuthorizations),
    TrafficPolicyAlreadyExists(TrafficPolicyAlreadyExists),
    TrafficPolicyInUse(TrafficPolicyInUse),
    TrafficPolicyInstanceAlreadyExists(TrafficPolicyInstanceAlreadyExists),
    VpcAssociationAuthorizationNotFound(VpcAssociationAuthorizationNotFound),
    VpcAssociationNotFound(VpcAssociationNotFound),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ConcurrentModification(ConcurrentModification)

Another user submitted a request to create, update, or delete the object at the same time that you did. Retry the request.

Tuple Fields of ConcurrentModification

0: ConcurrentModification

ConflictingDomainExists(ConflictingDomainExists)

The cause of this error depends on the operation that you're performing:

Create a public hosted zone: Two hosted zones that have the same name or that have a parent/child relationship (example.com and test.example.com) can't have any common name servers. You tried to create a hosted zone that has the same name as an existing hosted zone or that's the parent or child of an existing hosted zone, and you specified a delegation set that shares one or more name servers with the existing hosted zone. For more information, see CreateReusableDelegationSet.

Create a private hosted zone: A hosted zone with the specified name already exists and is already associated with the Amazon VPC that you specified.

Associate VPCs with a private hosted zone: The VPC that you specified is already associated with another hosted zone that has the same name.

Tuple Fields of ConflictingDomainExists

0: ConflictingDomainExists

ConflictingTypes(ConflictingTypes)

You tried to update a traffic policy instance by using a traffic policy version that has a different DNS type than the current type for the instance. You specified the type in the JSON document in the CreateTrafficPolicy or CreateTrafficPolicyVersionrequest.

Tuple Fields of ConflictingTypes

0: ConflictingTypes

DnssecNotFound(DnssecNotFound)

The hosted zone doesn't have any DNSSEC resources.

Tuple Fields of DnssecNotFound

0: DnssecNotFound

DelegationSetAlreadyCreated(DelegationSetAlreadyCreated)

A delegation set with the same owner and caller reference combination has already been created.

Tuple Fields of DelegationSetAlreadyCreated

0: DelegationSetAlreadyCreated

DelegationSetAlreadyReusable(DelegationSetAlreadyReusable)

The specified delegation set has already been marked as reusable.

Tuple Fields of DelegationSetAlreadyReusable

0: DelegationSetAlreadyReusable

DelegationSetInUse(DelegationSetInUse)

The specified delegation contains associated hosted zones which must be deleted before the reusable delegation set can be deleted.

Tuple Fields of DelegationSetInUse

0: DelegationSetInUse

DelegationSetNotAvailable(DelegationSetNotAvailable)

You can create a hosted zone that has the same name as an existing hosted zone (example.com is common), but there is a limit to the number of hosted zones that have the same name. If you get this error, Amazon Route 53 has reached that limit. If you own the domain name and Route 53 generates this error, contact Customer Support.

Tuple Fields of DelegationSetNotAvailable

0: DelegationSetNotAvailable

DelegationSetNotReusable(DelegationSetNotReusable)

A reusable delegation set with the specified ID does not exist.

Tuple Fields of DelegationSetNotReusable

0: DelegationSetNotReusable

HealthCheckAlreadyExists(HealthCheckAlreadyExists)

The health check you're attempting to create already exists. Amazon Route 53 returns this error when you submit a request that has the following values:

The same value for CallerReference as an existing health check, and one or more values that differ from the existing health check that has the same caller reference.

The same value for CallerReference as a health check that you created and later deleted, regardless of the other settings in the request.

Tuple Fields of HealthCheckAlreadyExists

0: HealthCheckAlreadyExists

HealthCheckInUse(HealthCheckInUse)

This error code is not in use.

Tuple Fields of HealthCheckInUse

0: HealthCheckInUse

HealthCheckVersionMismatch(HealthCheckVersionMismatch)

The value of HealthCheckVersion in the request doesn't match the value of HealthCheckVersion in the health check.

Tuple Fields of HealthCheckVersionMismatch

0: HealthCheckVersionMismatch

HostedZoneAlreadyExists(HostedZoneAlreadyExists)

The hosted zone you're trying to create already exists. Amazon Route 53 returns this error when a hosted zone has already been created with the specified CallerReference.

Tuple Fields of HostedZoneAlreadyExists

0: HostedZoneAlreadyExists

HostedZoneNotEmpty(HostedZoneNotEmpty)

The hosted zone contains resource records that are not SOA or NS records.

Tuple Fields of HostedZoneNotEmpty

0: HostedZoneNotEmpty

HostedZoneNotFound(HostedZoneNotFound)

The specified HostedZone can't be found.

Tuple Fields of HostedZoneNotFound

0: HostedZoneNotFound

HostedZoneNotPrivate(HostedZoneNotPrivate)

The specified hosted zone is a public hosted zone, not a private hosted zone.

Tuple Fields of HostedZoneNotPrivate

0: HostedZoneNotPrivate

HostedZonePartiallyDelegated(HostedZonePartiallyDelegated)

The hosted zone nameservers don't match the parent nameservers. The hosted zone and parent must have the same nameservers.

Tuple Fields of HostedZonePartiallyDelegated

0: HostedZonePartiallyDelegated

IncompatibleVersion(IncompatibleVersion)

The resource you're trying to access is unsupported on this Amazon Route 53 endpoint.

Tuple Fields of IncompatibleVersion

0: IncompatibleVersion

InsufficientCloudWatchLogsResourcePolicy(InsufficientCloudWatchLogsResourcePolicy)

Amazon Route 53 doesn't have the permissions required to create log streams and send query logs to log streams. Possible causes include the following:

There is no resource policy that specifies the log group ARN in the value for Resource.

The resource policy that includes the log group ARN in the value for Resource doesn't have the necessary permissions.

The resource policy hasn't finished propagating yet.

The Key management service (KMS) key you specified doesn’t exist or it can’t be used with the log group associated with query log. Update or provide a resource policy to grant permissions for the KMS key.

Tuple Fields of InsufficientCloudWatchLogsResourcePolicy

0: InsufficientCloudWatchLogsResourcePolicy

InvalidArgument(InvalidArgument)

Parameter name is not valid.

Tuple Fields of InvalidArgument

0: InvalidArgument

InvalidChangeBatch(InvalidChangeBatch)

This exception contains a list of messages that might contain one or more error messages. Each error message indicates one error in the change batch.

Tuple Fields of InvalidChangeBatch

0: InvalidChangeBatch

InvalidDomainName(InvalidDomainName)

The specified domain name is not valid.

Tuple Fields of InvalidDomainName

0: InvalidDomainName

InvalidInput(InvalidInput)

The input is not valid.

Tuple Fields of InvalidInput

0: InvalidInput

InvalidKmsArn(InvalidKmsArn)

The KeyManagementServiceArn that you specified isn't valid to use with DNSSEC signing.

Tuple Fields of InvalidKmsArn

0: InvalidKmsArn

InvalidKeySigningKeyName(InvalidKeySigningKeyName)

The key-signing key (KSK) name that you specified isn't a valid name.

Tuple Fields of InvalidKeySigningKeyName

0: InvalidKeySigningKeyName

InvalidKeySigningKeyStatus(InvalidKeySigningKeyStatus)

The key-signing key (KSK) status isn't valid or another KSK has the status INTERNAL_FAILURE.

Tuple Fields of InvalidKeySigningKeyStatus

0: InvalidKeySigningKeyStatus

InvalidPaginationToken(InvalidPaginationToken)

The value that you specified to get the second or subsequent page of results is invalid.

Tuple Fields of InvalidPaginationToken

0: InvalidPaginationToken

InvalidSigningStatus(InvalidSigningStatus)

Your hosted zone status isn't valid for this operation. In the hosted zone, change the status to enable DNSSEC or disable DNSSEC.

Tuple Fields of InvalidSigningStatus

0: InvalidSigningStatus

InvalidTrafficPolicyDocument(InvalidTrafficPolicyDocument)

The format of the traffic policy document that you specified in the Document element is not valid.

Tuple Fields of InvalidTrafficPolicyDocument

0: InvalidTrafficPolicyDocument

InvalidVpcId(InvalidVpcId)

The VPC ID that you specified either isn't a valid ID or the current account is not authorized to access this VPC.

Tuple Fields of InvalidVpcId

0: InvalidVpcId

KeySigningKeyAlreadyExists(KeySigningKeyAlreadyExists)

You've already created a key-signing key (KSK) with this name or with the same customer managed customer master key (CMK) ARN.

Tuple Fields of KeySigningKeyAlreadyExists

0: KeySigningKeyAlreadyExists

KeySigningKeyInParentDsRecord(KeySigningKeyInParentDsRecord)

The key-signing key (KSK) is specified in a parent DS record.

Tuple Fields of KeySigningKeyInParentDsRecord

0: KeySigningKeyInParentDsRecord

KeySigningKeyInUse(KeySigningKeyInUse)

The key-signing key (KSK) that you specified can't be deactivated because it's the only KSK for a currently-enabled DNSSEC. Disable DNSSEC signing, or add or enable another KSK.

Tuple Fields of KeySigningKeyInUse

0: KeySigningKeyInUse

KeySigningKeyWithActiveStatusNotFound(KeySigningKeyWithActiveStatusNotFound)

A key-signing key (KSK) with ACTIVE status wasn't found.

Tuple Fields of KeySigningKeyWithActiveStatusNotFound

0: KeySigningKeyWithActiveStatusNotFound

LastVpcAssociation(LastVpcAssociation)

The VPC that you're trying to disassociate from the private hosted zone is the last VPC that is associated with the hosted zone. Amazon Route 53 doesn't support disassociating the last VPC from a hosted zone.

Tuple Fields of LastVpcAssociation

0: LastVpcAssociation

LimitsExceeded(LimitsExceeded)

This operation can't be completed either because the current account has reached the limit on reusable delegation sets that it can create or because you've reached the limit on the number of Amazon VPCs that you can associate with a private hosted zone. To get the current limit on the number of reusable delegation sets, see GetAccountLimit. To get the current limit on the number of Amazon VPCs that you can associate with a private hosted zone, see GetHostedZoneLimit. To request a higher limit, create a case with the Amazon Web Services Support Center.

Tuple Fields of LimitsExceeded

0: LimitsExceeded

NoSuchChange(NoSuchChange)

A change with the specified change ID does not exist.

Tuple Fields of NoSuchChange

0: NoSuchChange

NoSuchCloudWatchLogsLogGroup(NoSuchCloudWatchLogsLogGroup)

There is no CloudWatch Logs log group with the specified ARN.

Tuple Fields of NoSuchCloudWatchLogsLogGroup

0: NoSuchCloudWatchLogsLogGroup

NoSuchDelegationSet(NoSuchDelegationSet)

A reusable delegation set with the specified ID does not exist.

Tuple Fields of NoSuchDelegationSet

0: NoSuchDelegationSet

NoSuchGeoLocation(NoSuchGeoLocation)

Amazon Route 53 doesn't support the specified geographic location. For a list of supported geolocation codes, see the GeoLocation data type.

Tuple Fields of NoSuchGeoLocation

0: NoSuchGeoLocation

NoSuchHealthCheck(NoSuchHealthCheck)

No health check exists with the specified ID.

Tuple Fields of NoSuchHealthCheck

0: NoSuchHealthCheck

NoSuchHostedZone(NoSuchHostedZone)

No hosted zone exists with the ID that you specified.

Tuple Fields of NoSuchHostedZone

0: NoSuchHostedZone

NoSuchKeySigningKey(NoSuchKeySigningKey)

The specified key-signing key (KSK) doesn't exist.

Tuple Fields of NoSuchKeySigningKey

0: NoSuchKeySigningKey

NoSuchQueryLoggingConfig(NoSuchQueryLoggingConfig)

There is no DNS query logging configuration with the specified ID.

Tuple Fields of NoSuchQueryLoggingConfig

0: NoSuchQueryLoggingConfig

NoSuchTrafficPolicy(NoSuchTrafficPolicy)

No traffic policy exists with the specified ID.

Tuple Fields of NoSuchTrafficPolicy

0: NoSuchTrafficPolicy

NoSuchTrafficPolicyInstance(NoSuchTrafficPolicyInstance)

No traffic policy instance exists with the specified ID.

Tuple Fields of NoSuchTrafficPolicyInstance

0: NoSuchTrafficPolicyInstance

NotAuthorizedException(NotAuthorizedException)

Associating the specified VPC with the specified hosted zone has not been authorized.

Tuple Fields of NotAuthorizedException

0: NotAuthorizedException

PriorRequestNotComplete(PriorRequestNotComplete)

If Amazon Route 53 can't process a request before the next request arrives, it will reject subsequent requests for the same hosted zone and return an HTTP 400 error (Bad request). If Route 53 returns this error repeatedly for the same request, we recommend that you wait, in intervals of increasing duration, before you try the request again.

Tuple Fields of PriorRequestNotComplete

0: PriorRequestNotComplete

PublicZoneVpcAssociation(PublicZoneVpcAssociation)

You're trying to associate a VPC with a public hosted zone. Amazon Route 53 doesn't support associating a VPC with a public hosted zone.

Tuple Fields of PublicZoneVpcAssociation

0: PublicZoneVpcAssociation

QueryLoggingConfigAlreadyExists(QueryLoggingConfigAlreadyExists)

You can create only one query logging configuration for a hosted zone, and a query logging configuration already exists for this hosted zone.

Tuple Fields of QueryLoggingConfigAlreadyExists

0: QueryLoggingConfigAlreadyExists

ThrottlingException(ThrottlingException)

The limit on the number of requests per second was exceeded.

Tuple Fields of ThrottlingException

0: ThrottlingException

TooManyHealthChecks(TooManyHealthChecks)

This health check can't be created because the current account has reached the limit on the number of active health checks.

For information about default limits, see Limits in the Amazon Route 53 Developer Guide.

For information about how to get the current limit for an account, see GetAccountLimit. To request a higher limit, create a case with the Amazon Web Services Support Center.

You have reached the maximum number of active health checks for an Amazon Web Services account. To request a higher limit, create a case with the Amazon Web Services Support Center.

Tuple Fields of TooManyHealthChecks

0: TooManyHealthChecks

TooManyHostedZones(TooManyHostedZones)

This operation can't be completed either because the current account has reached the limit on the number of hosted zones or because you've reached the limit on the number of hosted zones that can be associated with a reusable delegation set.

For information about default limits, see Limits in the Amazon Route 53 Developer Guide.

To get the current limit on hosted zones that can be created by an account, see GetAccountLimit.

To get the current limit on hosted zones that can be associated with a reusable delegation set, see GetReusableDelegationSetLimit.

To request a higher limit, create a case with the Amazon Web Services Support Center.

Tuple Fields of TooManyHostedZones

0: TooManyHostedZones

TooManyKeySigningKeys(TooManyKeySigningKeys)

You've reached the limit for the number of key-signing keys (KSKs). Remove at least one KSK, and then try again.

Tuple Fields of TooManyKeySigningKeys

0: TooManyKeySigningKeys

TooManyTrafficPolicies(TooManyTrafficPolicies)

This traffic policy can't be created because the current account has reached the limit on the number of traffic policies.

For information about default limits, see Limits in the Amazon Route 53 Developer Guide.

To get the current limit for an account, see GetAccountLimit.

To request a higher limit, create a case with the Amazon Web Services Support Center.

Tuple Fields of TooManyTrafficPolicies

0: TooManyTrafficPolicies

TooManyTrafficPolicyInstances(TooManyTrafficPolicyInstances)

This traffic policy instance can't be created because the current account has reached the limit on the number of traffic policy instances.

For information about default limits, see Limits in the Amazon Route 53 Developer Guide.

For information about how to get the current limit for an account, see GetAccountLimit.

To request a higher limit, create a case with the Amazon Web Services Support Center.

Tuple Fields of TooManyTrafficPolicyInstances

0: TooManyTrafficPolicyInstances

TooManyTrafficPolicyVersionsForCurrentPolicy(TooManyTrafficPolicyVersionsForCurrentPolicy)

This traffic policy version can't be created because you've reached the limit of 1000 on the number of versions that you can create for the current traffic policy.

To create more traffic policy versions, you can use GetTrafficPolicy to get the traffic policy document for a specified traffic policy version, and then use CreateTrafficPolicy to create a new traffic policy using the traffic policy document.

Tuple Fields of TooManyTrafficPolicyVersionsForCurrentPolicy

0: TooManyTrafficPolicyVersionsForCurrentPolicy

TooManyVpcAssociationAuthorizations(TooManyVpcAssociationAuthorizations)

You've created the maximum number of authorizations that can be created for the specified hosted zone. To authorize another VPC to be associated with the hosted zone, submit a DeleteVPCAssociationAuthorization request to remove an existing authorization. To get a list of existing authorizations, submit a ListVPCAssociationAuthorizations request.

Tuple Fields of TooManyVpcAssociationAuthorizations

0: TooManyVpcAssociationAuthorizations

TrafficPolicyAlreadyExists(TrafficPolicyAlreadyExists)

A traffic policy that has the same value for Name already exists.

Tuple Fields of TrafficPolicyAlreadyExists

0: TrafficPolicyAlreadyExists

TrafficPolicyInUse(TrafficPolicyInUse)

One or more traffic policy instances were created by using the specified traffic policy.

Tuple Fields of TrafficPolicyInUse

0: TrafficPolicyInUse

TrafficPolicyInstanceAlreadyExists(TrafficPolicyInstanceAlreadyExists)

There is already a traffic policy instance with the specified ID.

Tuple Fields of TrafficPolicyInstanceAlreadyExists

0: TrafficPolicyInstanceAlreadyExists

VpcAssociationAuthorizationNotFound(VpcAssociationAuthorizationNotFound)

The VPC that you specified is not authorized to be associated with the hosted zone.

Tuple Fields of VpcAssociationAuthorizationNotFound

0: VpcAssociationAuthorizationNotFound

VpcAssociationNotFound(VpcAssociationNotFound)

The specified VPC and hosted zone are not currently associated.

Tuple Fields of VpcAssociationNotFound

0: VpcAssociationNotFound

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>