Module aws_sdk_route53::error

Modules

concurrent_modification

See ConcurrentModification

conflicting_domain_exists

See ConflictingDomainExists

conflicting_types

See ConflictingTypes

delegation_set_already_created

See DelegationSetAlreadyCreated

delegation_set_already_reusable

See DelegationSetAlreadyReusable

delegation_set_in_use

See DelegationSetInUse

delegation_set_not_available

See DelegationSetNotAvailable

delegation_set_not_reusable

See DelegationSetNotReusable

dnssec_not_found

See DnssecNotFound

health_check_already_exists

See HealthCheckAlreadyExists

health_check_in_use

See HealthCheckInUse

health_check_version_mismatch

See HealthCheckVersionMismatch

hosted_zone_already_exists

See HostedZoneAlreadyExists

hosted_zone_not_empty

See HostedZoneNotEmpty

hosted_zone_not_found

See HostedZoneNotFound

hosted_zone_not_private

See HostedZoneNotPrivate

hosted_zone_partially_delegated

See HostedZonePartiallyDelegated

incompatible_version

See IncompatibleVersion

insufficient_cloud_watch_logs_resource_policy

See InsufficientCloudWatchLogsResourcePolicy

invalid_argument

See InvalidArgument

invalid_change_batch

See InvalidChangeBatch

invalid_domain_name

See InvalidDomainName

invalid_input

See InvalidInput

invalid_key_signing_key_name

See InvalidKeySigningKeyName

invalid_key_signing_key_status

See InvalidKeySigningKeyStatus

invalid_kms_arn

See InvalidKmsArn

invalid_pagination_token

See InvalidPaginationToken

invalid_signing_status

See InvalidSigningStatus

invalid_traffic_policy_document

See InvalidTrafficPolicyDocument

invalid_vpc_id

See InvalidVpcId

key_signing_key_already_exists

See KeySigningKeyAlreadyExists

key_signing_key_in_parent_ds_record

See KeySigningKeyInParentDsRecord

key_signing_key_in_use

See KeySigningKeyInUse

key_signing_key_with_active_status_not_found

See KeySigningKeyWithActiveStatusNotFound

last_vpc_association

See LastVpcAssociation

limits_exceeded

See LimitsExceeded

no_such_change

See NoSuchChange

no_such_cloud_watch_logs_log_group

See NoSuchCloudWatchLogsLogGroup

no_such_delegation_set

See NoSuchDelegationSet

no_such_geo_location

See NoSuchGeoLocation

no_such_health_check

See NoSuchHealthCheck

no_such_hosted_zone

See NoSuchHostedZone

no_such_key_signing_key

See NoSuchKeySigningKey

no_such_query_logging_config

See NoSuchQueryLoggingConfig

no_such_traffic_policy

See NoSuchTrafficPolicy

no_such_traffic_policy_instance

See NoSuchTrafficPolicyInstance

not_authorized_exception

See NotAuthorizedException

prior_request_not_complete

See PriorRequestNotComplete

public_zone_vpc_association

See PublicZoneVpcAssociation

query_logging_config_already_exists

See QueryLoggingConfigAlreadyExists

throttling_exception

See ThrottlingException

too_many_health_checks

See TooManyHealthChecks

too_many_hosted_zones

See TooManyHostedZones

too_many_key_signing_keys

See TooManyKeySigningKeys

too_many_traffic_policies

See TooManyTrafficPolicies

too_many_traffic_policy_instances

See TooManyTrafficPolicyInstances

too_many_traffic_policy_versions_for_current_policy

See TooManyTrafficPolicyVersionsForCurrentPolicy

too_many_vpc_association_authorizations

See TooManyVpcAssociationAuthorizations

traffic_policy_already_exists

See TrafficPolicyAlreadyExists

traffic_policy_in_use

See TrafficPolicyInUse

traffic_policy_instance_already_exists

See TrafficPolicyInstanceAlreadyExists

vpc_association_authorization_not_found

See VpcAssociationAuthorizationNotFound

vpc_association_not_found

See VpcAssociationNotFound

Structs

ActivateKeySigningKeyError

AssociateVPCWithHostedZoneError

ChangeResourceRecordSetsError

ChangeTagsForResourceError

ConcurrentModification

Another user submitted a request to create, update, or delete the object at the same time that you did. Retry the request.

ConflictingDomainExists

The cause of this error depends on the operation that you're performing:

ConflictingTypes

You tried to update a traffic policy instance by using a traffic policy version that has a different DNS type

CreateHealthCheckError

CreateHostedZoneError

CreateKeySigningKeyError

CreateQueryLoggingConfigError

CreateReusableDelegationSetError

CreateTrafficPolicyError

CreateTrafficPolicyInstanceError

CreateTrafficPolicyVersionError

CreateVPCAssociationAuthorizationError

DeactivateKeySigningKeyError

DelegationSetAlreadyCreated

A delegation set with the same owner and caller reference combination has already been created.

DelegationSetAlreadyReusable

The specified delegation set has already been marked as reusable.

DelegationSetInUse

The specified delegation contains associated hosted zones which must be deleted before the reusable delegation set

DelegationSetNotAvailable

You can create a hosted zone that has the same name as an existing hosted zone (example.com is common), but there is a

DelegationSetNotReusable

A reusable delegation set with the specified ID does not exist.

DeleteHealthCheckError

DeleteHostedZoneError

DeleteKeySigningKeyError

DeleteQueryLoggingConfigError

DeleteReusableDelegationSetError

DeleteTrafficPolicyError

DeleteTrafficPolicyInstanceError

DeleteVPCAssociationAuthorizationError

DisableHostedZoneDNSSECError

DisassociateVPCFromHostedZoneError

DnssecNotFound

The hosted zone doesn't have any DNSSEC resources.

EnableHostedZoneDNSSECError

GetAccountLimitError

GetChangeError

GetCheckerIpRangesError

GetDNSSECError

GetGeoLocationError

GetHealthCheckCountError

GetHealthCheckError

GetHealthCheckLastFailureReasonError

GetHealthCheckStatusError

GetHostedZoneCountError

GetHostedZoneError

GetHostedZoneLimitError

GetQueryLoggingConfigError

GetReusableDelegationSetError

GetReusableDelegationSetLimitError

GetTrafficPolicyError

GetTrafficPolicyInstanceCountError

GetTrafficPolicyInstanceError

HealthCheckAlreadyExists

The health check you're attempting to create already exists. Amazon Route 53 returns this error when you submit a request that

HealthCheckInUse

This error code is not in use.

HealthCheckVersionMismatch

The value of HealthCheckVersion in the request doesn't match the value of HealthCheckVersion

HostedZoneAlreadyExists

The hosted zone you're trying to create already exists. Amazon Route 53 returns this error when a hosted zone has already been created

HostedZoneNotEmpty

The hosted zone contains resource records that are not SOA or NS records.

HostedZoneNotFound

The specified HostedZone can't be found.

HostedZoneNotPrivate

The specified hosted zone is a public hosted zone, not a private hosted zone.

HostedZonePartiallyDelegated

The hosted zone nameservers don't match the parent nameservers. The hosted zone and parent must have the same nameservers.

IncompatibleVersion

The resource you're trying to access is unsupported on this Amazon Route 53 endpoint.

InsufficientCloudWatchLogsResourcePolicy

Amazon Route 53 doesn't have the permissions required to create log streams and send query logs to log streams. Possible causes

InvalidArgument

Parameter name is not valid.

InvalidChangeBatch

This exception contains a list of messages that might contain one or more error messages. Each error message indicates

InvalidDomainName

The specified domain name is not valid.

InvalidInput

The input is not valid.

InvalidKeySigningKeyName

The key-signing key (KSK) name that you specified isn't a valid name.

InvalidKeySigningKeyStatus

The key-signing key (KSK) status isn't valid or another KSK has the status INTERNAL_FAILURE.

InvalidKmsArn

The KeyManagementServiceArn that you specified isn't valid to use with DNSSEC signing.

InvalidPaginationToken

The value that you specified to get the second or subsequent page of results is invalid.

InvalidSigningStatus

Your hosted zone status isn't valid for this operation. In the hosted zone, change the status to enable DNSSEC

InvalidTrafficPolicyDocument

The format of the traffic policy document that you specified in the Document element is not valid.

InvalidVpcId

The VPC ID that you specified either isn't a valid ID or the current account is not authorized to access this VPC.

KeySigningKeyAlreadyExists

You've already created a key-signing key (KSK) with this name or with the same customer managed customer master key (CMK) ARN.

KeySigningKeyInParentDsRecord

The key-signing key (KSK) is specified in a parent DS record.

KeySigningKeyInUse

The key-signing key (KSK) that you specified can't be deactivated because it's the only KSK for a

KeySigningKeyWithActiveStatusNotFound

A key-signing key (KSK) with ACTIVE status wasn't found.

LastVpcAssociation

The VPC that you're trying to disassociate from the private hosted zone is the last VPC that is associated with

LimitsExceeded

This operation can't be completed either because the current account has reached the limit on reusable delegation sets

ListGeoLocationsError

ListHealthChecksError

ListHostedZonesByNameError

ListHostedZonesByVPCError

ListHostedZonesError

ListQueryLoggingConfigsError

ListResourceRecordSetsError

ListReusableDelegationSetsError

ListTagsForResourceError

ListTagsForResourcesError

ListTrafficPoliciesError

ListTrafficPolicyInstancesByHostedZoneError

ListTrafficPolicyInstancesByPolicyError

ListTrafficPolicyInstancesError

ListTrafficPolicyVersionsError

ListVPCAssociationAuthorizationsError

NoSuchChange

A change with the specified change ID does not exist.

NoSuchCloudWatchLogsLogGroup

There is no CloudWatch Logs log group with the specified ARN.

NoSuchDelegationSet

A reusable delegation set with the specified ID does not exist.

NoSuchGeoLocation

Amazon Route 53 doesn't support the specified geographic location. For a list of supported geolocation codes, see the

NoSuchHealthCheck

No health check exists with the specified ID.

NoSuchHostedZone

No hosted zone exists with the ID that you specified.

NoSuchKeySigningKey

The specified key-signing key (KSK) doesn't exist.

NoSuchQueryLoggingConfig

There is no DNS query logging configuration with the specified ID.

NoSuchTrafficPolicy

No traffic policy exists with the specified ID.

NoSuchTrafficPolicyInstance

No traffic policy instance exists with the specified ID.

NotAuthorizedException

Associating the specified VPC with the specified hosted zone has not been authorized.

PriorRequestNotComplete

If Amazon Route 53 can't process a request before the next request arrives, it will reject

PublicZoneVpcAssociation

You're trying to associate a VPC with a public hosted zone. Amazon Route 53 doesn't support associating a

QueryLoggingConfigAlreadyExists

You can create only one query logging configuration for a hosted zone, and a query logging configuration already exists

TestDNSAnswerError

ThrottlingException

The limit on the number of requests per second was exceeded.

TooManyHealthChecks

This health check can't be created because the current account has reached the limit on the number of active health checks.

TooManyHostedZones

This operation can't be completed either because the current account has reached the limit on the number of hosted zones

TooManyKeySigningKeys

You've reached the limit for the number of key-signing keys (KSKs). Remove at least one KSK, and then try again.

TooManyTrafficPolicies

This traffic policy can't be created because the current account has reached the limit on the number of traffic policies.

TooManyTrafficPolicyInstances

This traffic policy instance can't be created because the current account has reached the limit on the number of

TooManyTrafficPolicyVersionsForCurrentPolicy

This traffic policy version can't be created because you've reached the limit of 1000 on the number of versions

TooManyVpcAssociationAuthorizations

You've created the maximum number of authorizations that can be created for the specified hosted zone.

TrafficPolicyAlreadyExists

A traffic policy that has the same value for Name already exists.

TrafficPolicyInUse

One or more traffic policy instances were created by using the specified traffic policy.

TrafficPolicyInstanceAlreadyExists

There is already a traffic policy instance with the specified ID.

UpdateHealthCheckError

UpdateHostedZoneCommentError

UpdateTrafficPolicyCommentError

UpdateTrafficPolicyInstanceError

VpcAssociationAuthorizationNotFound

The VPC that you specified is not authorized to be associated with the hosted zone.

VpcAssociationNotFound

The specified VPC and hosted zone are not currently associated.

Enums

ActivateKeySigningKeyErrorKind

AssociateVPCWithHostedZoneErrorKind

ChangeResourceRecordSetsErrorKind

ChangeTagsForResourceErrorKind

CreateHealthCheckErrorKind

CreateHostedZoneErrorKind

CreateKeySigningKeyErrorKind

CreateQueryLoggingConfigErrorKind

CreateReusableDelegationSetErrorKind

CreateTrafficPolicyErrorKind

CreateTrafficPolicyInstanceErrorKind

CreateTrafficPolicyVersionErrorKind

CreateVPCAssociationAuthorizationErrorKind

DeactivateKeySigningKeyErrorKind

DeleteHealthCheckErrorKind

DeleteHostedZoneErrorKind

DeleteKeySigningKeyErrorKind

DeleteQueryLoggingConfigErrorKind

DeleteReusableDelegationSetErrorKind

DeleteTrafficPolicyErrorKind

DeleteTrafficPolicyInstanceErrorKind

DeleteVPCAssociationAuthorizationErrorKind

DisableHostedZoneDNSSECErrorKind

DisassociateVPCFromHostedZoneErrorKind

EnableHostedZoneDNSSECErrorKind

GetAccountLimitErrorKind

GetChangeErrorKind

GetCheckerIpRangesErrorKind

GetDNSSECErrorKind

GetGeoLocationErrorKind

GetHealthCheckCountErrorKind

GetHealthCheckErrorKind

GetHealthCheckLastFailureReasonErrorKind

GetHealthCheckStatusErrorKind

GetHostedZoneCountErrorKind

GetHostedZoneErrorKind

GetHostedZoneLimitErrorKind

GetQueryLoggingConfigErrorKind

GetReusableDelegationSetErrorKind

GetReusableDelegationSetLimitErrorKind

GetTrafficPolicyErrorKind

GetTrafficPolicyInstanceCountErrorKind

GetTrafficPolicyInstanceErrorKind

ListGeoLocationsErrorKind

ListHealthChecksErrorKind

ListHostedZonesByNameErrorKind

ListHostedZonesByVPCErrorKind

ListHostedZonesErrorKind

ListQueryLoggingConfigsErrorKind

ListResourceRecordSetsErrorKind

ListReusableDelegationSetsErrorKind

ListTagsForResourceErrorKind

ListTagsForResourcesErrorKind

ListTrafficPoliciesErrorKind

ListTrafficPolicyInstancesByHostedZoneErrorKind

ListTrafficPolicyInstancesByPolicyErrorKind

ListTrafficPolicyInstancesErrorKind

ListTrafficPolicyVersionsErrorKind

ListVPCAssociationAuthorizationsErrorKind

TestDNSAnswerErrorKind

UpdateHealthCheckErrorKind

UpdateHostedZoneCommentErrorKind

UpdateTrafficPolicyCommentErrorKind

UpdateTrafficPolicyInstanceErrorKind