Module aws_sdk_route53::input

Modules

activate_key_signing_key_input

See ActivateKeySigningKeyInput

associate_vpc_with_hosted_zone_input

See AssociateVpcWithHostedZoneInput

change_resource_record_sets_input

See ChangeResourceRecordSetsInput

change_tags_for_resource_input

See ChangeTagsForResourceInput

create_health_check_input

See CreateHealthCheckInput

create_hosted_zone_input

See CreateHostedZoneInput

create_key_signing_key_input

See CreateKeySigningKeyInput

create_query_logging_config_input

See CreateQueryLoggingConfigInput

create_reusable_delegation_set_input

See CreateReusableDelegationSetInput

create_traffic_policy_input

See CreateTrafficPolicyInput

create_traffic_policy_instance_input

See CreateTrafficPolicyInstanceInput

create_traffic_policy_version_input

See CreateTrafficPolicyVersionInput

create_vpc_association_authorization_input

See CreateVpcAssociationAuthorizationInput

deactivate_key_signing_key_input

See DeactivateKeySigningKeyInput

delete_health_check_input

See DeleteHealthCheckInput

delete_hosted_zone_input

See DeleteHostedZoneInput

delete_key_signing_key_input

See DeleteKeySigningKeyInput

delete_query_logging_config_input

See DeleteQueryLoggingConfigInput

delete_reusable_delegation_set_input

See DeleteReusableDelegationSetInput

delete_traffic_policy_input

See DeleteTrafficPolicyInput

delete_traffic_policy_instance_input

See DeleteTrafficPolicyInstanceInput

delete_vpc_association_authorization_input

See DeleteVpcAssociationAuthorizationInput

disable_hosted_zone_dnssec_input

See DisableHostedZoneDnssecInput

disassociate_vpc_from_hosted_zone_input

See DisassociateVpcFromHostedZoneInput

enable_hosted_zone_dnssec_input

See EnableHostedZoneDnssecInput

get_account_limit_input

See GetAccountLimitInput

get_change_input

See GetChangeInput

get_checker_ip_ranges_input

See GetCheckerIpRangesInput

get_dnssec_input

See GetDnssecInput

get_geo_location_input

See GetGeoLocationInput

get_health_check_count_input

See GetHealthCheckCountInput

get_health_check_input

See GetHealthCheckInput

get_health_check_last_failure_reason_input

See GetHealthCheckLastFailureReasonInput

get_health_check_status_input

See GetHealthCheckStatusInput

get_hosted_zone_count_input

See GetHostedZoneCountInput

get_hosted_zone_input

See GetHostedZoneInput

get_hosted_zone_limit_input

See GetHostedZoneLimitInput

get_query_logging_config_input

See GetQueryLoggingConfigInput

get_reusable_delegation_set_input

See GetReusableDelegationSetInput

get_reusable_delegation_set_limit_input

See GetReusableDelegationSetLimitInput

get_traffic_policy_input

See GetTrafficPolicyInput

get_traffic_policy_instance_count_input

See GetTrafficPolicyInstanceCountInput

get_traffic_policy_instance_input

See GetTrafficPolicyInstanceInput

list_geo_locations_input

See ListGeoLocationsInput

list_health_checks_input

See ListHealthChecksInput

list_hosted_zones_by_name_input

See ListHostedZonesByNameInput

list_hosted_zones_by_vpc_input

See ListHostedZonesByVpcInput

list_hosted_zones_input

See ListHostedZonesInput

list_query_logging_configs_input

See ListQueryLoggingConfigsInput

list_resource_record_sets_input

See ListResourceRecordSetsInput

list_reusable_delegation_sets_input

See ListReusableDelegationSetsInput

list_tags_for_resource_input

See ListTagsForResourceInput

list_tags_for_resources_input

See ListTagsForResourcesInput

list_traffic_policies_input

See ListTrafficPoliciesInput

list_traffic_policy_instances_by_hosted_zone_input

See ListTrafficPolicyInstancesByHostedZoneInput

list_traffic_policy_instances_by_policy_input

See ListTrafficPolicyInstancesByPolicyInput

list_traffic_policy_instances_input

See ListTrafficPolicyInstancesInput

list_traffic_policy_versions_input

See ListTrafficPolicyVersionsInput

list_vpc_association_authorizations_input

See ListVpcAssociationAuthorizationsInput

test_dns_answer_input

See TestDnsAnswerInput

update_health_check_input

See UpdateHealthCheckInput

update_hosted_zone_comment_input

See UpdateHostedZoneCommentInput

update_traffic_policy_comment_input

See UpdateTrafficPolicyCommentInput

update_traffic_policy_instance_input

See UpdateTrafficPolicyInstanceInput

Structs

ActivateKeySigningKeyInput

AssociateVpcWithHostedZoneInput

A complex type that contains information about the request to associate a VPC with a private hosted zone.

ChangeResourceRecordSetsInput

A complex type that contains change information for the resource record set.

ChangeTagsForResourceInput

A complex type that contains information about the tags that you want to add, edit, or delete.

CreateHealthCheckInput

A complex type that contains the health check request information.

CreateHostedZoneInput

A complex type that contains information about the request to create a public or private hosted zone.

CreateKeySigningKeyInput

CreateQueryLoggingConfigInput

CreateReusableDelegationSetInput

CreateTrafficPolicyInput

A complex type that contains information about the traffic policy that you want to create.

CreateTrafficPolicyInstanceInput

A complex type that contains information about the resource record sets that you want to create based on a specified traffic policy.

CreateTrafficPolicyVersionInput

A complex type that contains information about the traffic policy that you want to create a new version for.

CreateVpcAssociationAuthorizationInput

A complex type that contains information about the request to authorize associating a VPC with your private hosted zone.

DeactivateKeySigningKeyInput

DeleteHealthCheckInput

This action deletes a health check.

DeleteHostedZoneInput

A request to delete a hosted zone.

DeleteKeySigningKeyInput

DeleteQueryLoggingConfigInput

DeleteReusableDelegationSetInput

A request to delete a reusable delegation set.

DeleteTrafficPolicyInput

A request to delete a specified traffic policy version.

DeleteTrafficPolicyInstanceInput

A request to delete a specified traffic policy instance.

DeleteVpcAssociationAuthorizationInput

A complex type that contains information about the request to remove authorization to associate a VPC

DisableHostedZoneDnssecInput

DisassociateVpcFromHostedZoneInput

A complex type that contains information about the VPC that you want to disassociate from a

EnableHostedZoneDnssecInput

GetAccountLimitInput

A complex type that contains information about the request to create a hosted zone.

GetChangeInput

The input for a GetChange request.

GetCheckerIpRangesInput

Empty request.

GetDnssecInput

GetGeoLocationInput

A request for information about whether a specified geographic location is supported for Amazon Route 53 geolocation resource record sets.

GetHealthCheckCountInput

A request for the number of health checks that are associated with the current AWS account.

GetHealthCheckInput

A request to get information about a specified health check.

GetHealthCheckLastFailureReasonInput

A request for the reason that a health check failed most recently.

GetHealthCheckStatusInput

A request to get the status for a health check.

GetHostedZoneCountInput

A request to retrieve a count of all the hosted zones that are associated with the current AWS account.

GetHostedZoneInput

A request to get information about a specified hosted zone.

GetHostedZoneLimitInput

A complex type that contains information about the request to create a hosted zone.

GetQueryLoggingConfigInput

GetReusableDelegationSetInput

A request to get information about a specified reusable delegation set.

GetReusableDelegationSetLimitInput

A complex type that contains information about the request to create a hosted zone.

GetTrafficPolicyInput

Gets information about a specific traffic policy version.

GetTrafficPolicyInstanceCountInput

Request to get the number of traffic policy instances that are associated with the current AWS account.

GetTrafficPolicyInstanceInput

Gets information about a specified traffic policy instance.

ListGeoLocationsInput

A request to get a list of geographic locations that Amazon Route 53 supports for geolocation resource record sets.

ListHealthChecksInput

A request to retrieve a list of the health checks that are associated with the current AWS account.

ListHostedZonesByNameInput

Retrieves a list of the public and private hosted zones that are associated with the current AWS account in ASCII order by domain

ListHostedZonesByVpcInput

Lists all the private hosted zones that a specified VPC is associated with, regardless of which AWS account created the hosted zones.

ListHostedZonesInput

A request to retrieve a list of the public and private hosted zones that are associated with the current AWS account.

ListQueryLoggingConfigsInput

ListResourceRecordSetsInput

A request for the resource record sets that are associated with a specified hosted zone.

ListReusableDelegationSetsInput

A request to get a list of the reusable delegation sets that are associated with the current AWS account.

ListTagsForResourceInput

A complex type containing information about a request for a list of the tags that are associated with an individual resource.

ListTagsForResourcesInput

A complex type that contains information about the health checks or hosted zones for which you want to list tags.

ListTrafficPoliciesInput

A complex type that contains the information about the request to list the traffic policies that are associated

ListTrafficPolicyInstancesByHostedZoneInput

A request for the traffic policy instances that you created in a specified hosted zone.

ListTrafficPolicyInstancesByPolicyInput

A complex type that contains the information about the request to list your traffic policy instances.

ListTrafficPolicyInstancesInput

A request to get information about the traffic policy instances that you created by using the current AWS account.

ListTrafficPolicyVersionsInput

A complex type that contains the information about the request to list your traffic

ListVpcAssociationAuthorizationsInput

A complex type that contains information about that can be associated with your hosted zone.

TestDnsAnswerInput

Gets the value that Amazon Route 53 returns in response to a DNS request for a specified record name and type. You can optionally specify

UpdateHealthCheckInput

A complex type that contains information about a request to update a health check.

UpdateHostedZoneCommentInput

A request to update the comment for a hosted zone.

UpdateTrafficPolicyCommentInput

A complex type that contains information about the traffic policy that you want to update the comment for.

UpdateTrafficPolicyInstanceInput

A complex type that contains information about the resource record sets that you want to update based on a specified traffic policy instance.