Module aws_sdk_route53::model

Modules

account_limit

See AccountLimit

alarm_identifier

See AlarmIdentifier

alias_target

See AliasTarget

change

See Change

change_batch

See ChangeBatch

change_info

See ChangeInfo

cloud_watch_alarm_configuration

See CloudWatchAlarmConfiguration

delegation_set

See DelegationSet

dimension

See Dimension

dnssec_status

See DnssecStatus

geo_location

See GeoLocation

geo_location_details

See GeoLocationDetails

health_check

See HealthCheck

health_check_config

See HealthCheckConfig

health_check_observation

See HealthCheckObservation

hosted_zone

See HostedZone

hosted_zone_config

See HostedZoneConfig

hosted_zone_limit

See HostedZoneLimit

hosted_zone_owner

See HostedZoneOwner

hosted_zone_summary

See HostedZoneSummary

key_signing_key

See KeySigningKey

linked_service

See LinkedService

query_logging_config

See QueryLoggingConfig

resource_record

See ResourceRecord

resource_record_set

See ResourceRecordSet

resource_tag_set

See ResourceTagSet

reusable_delegation_set_limit

See ReusableDelegationSetLimit

status_report

See StatusReport

tag

See Tag

traffic_policy

See TrafficPolicy

traffic_policy_instance

See TrafficPolicyInstance

traffic_policy_summary

See TrafficPolicySummary

vpc

See Vpc

Structs

AccountLimit

A complex type that contains the type of limit that you specified in the request and the current value for that limit.

AlarmIdentifier

A complex type that identifies the CloudWatch alarm that you want Amazon Route 53 health checkers to use to determine whether

AliasTarget

Change

The information for each resource record set that you want to change.

ChangeBatch

The information for a change request.

ChangeInfo

A complex type that describes change information about changes made to your hosted

CloudWatchAlarmConfiguration

A complex type that contains information about the CloudWatch alarm that Amazon Route 53 is monitoring for this health check.

DelegationSet

A complex type that lists the name servers in a delegation set, as well as the CallerReference and the ID

Dimension

For the metric that the CloudWatch alarm is associated with, a complex type that contains information about one dimension.

DnssecStatus

A string repesenting the status of DNSSEC signing.

GeoLocation

A complex type that contains information about a geographic location.

GeoLocationDetails

A complex type that contains the codes and full continent, country, and subdivision names for the specified geolocation code.

HealthCheck

A complex type that contains information about one health check that is associated with the current AWS account.

HealthCheckConfig

A complex type that contains information about the health check.

HealthCheckObservation

A complex type that contains the last failure reason as reported by one Amazon Route 53 health checker.

HostedZone

A complex type that contains general information about the hosted zone.

HostedZoneConfig

A complex type that contains an optional comment about your hosted zone. If you don't

HostedZoneLimit

A complex type that contains the type of limit that you specified in the request and the current value for that limit.

HostedZoneOwner

A complex type that identifies a hosted zone that a specified Amazon VPC is associated with and the owner of the hosted zone.

HostedZoneSummary

In the response to a ListHostedZonesByVPC request, the HostedZoneSummaries element contains one

KeySigningKey

A key-signing key (KSK) is a complex type that represents a public/private key pair. The private key is used to generate

LinkedService

If a health check or hosted zone was created by another service, LinkedService is a complex type that

QueryLoggingConfig

A complex type that contains information about a configuration for DNS query logging.

ResourceRecord

Information specific to the resource record.

ResourceRecordSet

Information about the resource record set to create or delete.

ResourceTagSet

A complex type containing a resource and its associated tags.

ReusableDelegationSetLimit

A complex type that contains the type of limit that you specified in the request and the current value for that limit.

StatusReport

A complex type that contains the status that one Amazon Route 53 health checker reports and the time of the health check.

Tag

A complex type that contains information about a tag that you want to add or edit for the specified health check or hosted zone.

TrafficPolicy

A complex type that contains settings for a traffic policy.

TrafficPolicyInstance

A complex type that contains settings for the new traffic policy instance.

TrafficPolicySummary

A complex type that contains information about the latest version of one traffic policy

Vpc

(Private hosted zones only) A complex type that contains information about an Amazon VPC.

Enums

AccountLimitType

ChangeAction

ChangeStatus

CloudWatchRegion

ComparisonOperator

HealthCheckRegion

HealthCheckType

HostedZoneLimitType

InsufficientDataHealthStatus

ResettableElementName

ResourceRecordSetFailover

ResourceRecordSetRegion

ReusableDelegationSetLimitType

RrType

Statistic

TagResourceType

VpcRegion