Struct aws_sdk_route53::model::key_signing_key::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for KeySigningKey

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

A string used to identify a key-signing key (KSK). Name can include numbers, letters, and underscores (_). Name must be unique for each key-signing key in the same hosted zone.

pub fn set_name(self, input: Option<String>) -> Self

pub fn kms_arn(self, input: impl Into<String>) -> Self

The Amazon resource name (ARN) used to identify the customer managed customer master key (CMK) in AWS Key Management Service (AWS KMS). The KmsArn must be unique for each key-signing key (KSK) in a single hosted zone.

You must configure the CMK as follows:

Status

Enabled

Key spec

ECC_NIST_P256

Key usage

Sign and verify

Key policy

The key policy must give permission for the following actions:

DescribeKey

GetPublicKey

Sign

The key policy must also include the Amazon Route 53 service in the principal for your account. Specify the following:

"Service": "api-service.dnssec.route53.aws.internal"

For more information about working with the customer managed CMK in AWS KMS, see AWS Key Management Service concepts.

pub fn set_kms_arn(self, input: Option<String>) -> Self

pub fn flag(self, input: i32) -> Self

An integer that specifies how the key is used. For key-signing key (KSK), this value is always 257.

pub fn set_flag(self, input: Option<i32>) -> Self

pub fn signing_algorithm_mnemonic(self, input: impl Into<String>) -> Self

A string used to represent the signing algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.1.

pub fn set_signing_algorithm_mnemonic(self, input: Option<String>) -> Self

pub fn signing_algorithm_type(self, input: i32) -> Self

An integer used to represent the signing algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.1.

pub fn set_signing_algorithm_type(self, input: Option<i32>) -> Self

pub fn digest_algorithm_mnemonic(self, input: impl Into<String>) -> Self

A string used to represent the delegation signer digest algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.3.

pub fn set_digest_algorithm_mnemonic(self, input: Option<String>) -> Self

pub fn digest_algorithm_type(self, input: i32) -> Self

An integer used to represent the delegation signer digest algorithm. This value must follow the guidelines provided by RFC-8624 Section 3.3.

pub fn set_digest_algorithm_type(self, input: Option<i32>) -> Self

pub fn key_tag(self, input: i32) -> Self

An integer used to identify the DNSSEC record for the domain name. The process used to calculate the value is described in RFC-4034 Appendix B.

pub fn set_key_tag(self, input: Option<i32>) -> Self

pub fn digest_value(self, input: impl Into<String>) -> Self

A cryptographic digest of a DNSKEY resource record (RR). DNSKEY records are used to publish the public key that resolvers can use to verify DNSSEC signatures that are used to secure certain kinds of information provided by the DNS system.

pub fn set_digest_value(self, input: Option<String>) -> Self

pub fn public_key(self, input: impl Into<String>) -> Self

The public key, represented as a Base64 encoding, as required by RFC-4034 Page 5.

pub fn set_public_key(self, input: Option<String>) -> Self

pub fn ds_record(self, input: impl Into<String>) -> Self

A string that represents a delegation signer (DS) record.

pub fn set_ds_record(self, input: Option<String>) -> Self

pub fn dnskey_record(self, input: impl Into<String>) -> Self

A string that represents a DNSKEY record.

pub fn set_dnskey_record(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

A string that represents the current key-signing key (KSK) status.

Status can have one of the following values:

ACTIVE

The KSK is being used for signing.

INACTIVE

The KSK is not being used for signing.

DELETING

The KSK is in the process of being deleted.

ACTION_NEEDED

There is a problem with the KSK that requires you to take action to resolve. For example, the customer managed customer master key (CMK) might have been deleted, or the permissions for the customer managed CMK might have been changed.

INTERNAL_FAILURE

There was an error during a request. Before you can continue to work with DNSSEC signing, including actions that involve this KSK, you must correct the problem. For example, you may need to activate or deactivate the KSK.

pub fn set_status(self, input: Option<String>) -> Self

pub fn status_message(self, input: impl Into<String>) -> Self

The status message provided for the following key-signing key (KSK) statuses: ACTION_NEEDED or INTERNAL_FAILURE. The status message includes information about what the problem might be and steps that you can take to correct the issue.

pub fn set_status_message(self, input: Option<String>) -> Self

pub fn created_date(self, input: Instant) -> Self

The date when the key-signing key (KSK) was created.

pub fn set_created_date(self, input: Option<Instant>) -> Self

pub fn last_modified_date(self, input: Instant) -> Self

The last time that the key-signing key (KSK) was changed.

pub fn set_last_modified_date(self, input: Option<Instant>) -> Self

pub fn build(self) -> KeySigningKey

Consumes the builder and constructs a KeySigningKey