Module aws_sdk_route53::operation

Structs

ActivateKeySigningKey

Activates a key-signing key (KSK) so that it can be used for signing by DNSSEC. This operation changes the

AssociateVPCWithHostedZone

Associates an Amazon VPC with a private hosted zone.

ChangeResourceRecordSets

Creates, changes, or deletes a resource record set, which contains authoritative DNS information for a specified

ChangeTagsForResource

Adds, edits, or deletes tags for a health check or a hosted zone.

CreateHealthCheck

Creates a new health check.

CreateHostedZone

Creates a new public or private hosted zone. You create records in a public hosted zone to define how you want to route traffic

CreateKeySigningKey

Creates a new key-signing key (KSK) associated with a hosted zone. You can only have two KSKs per hosted zone.

CreateQueryLoggingConfig

Creates a configuration for DNS query logging. After you create a query logging configuration, Amazon Route 53 begins to publish

CreateReusableDelegationSet

Creates a delegation set (a group of four name servers) that can be reused by multiple hosted zones that were created by

CreateTrafficPolicy

Creates a traffic policy, which you use to create multiple DNS resource record sets for one domain name (such as example.com) or

CreateTrafficPolicyInstance

Creates resource record sets in a specified hosted zone based on the settings in a specified traffic policy version.

CreateTrafficPolicyVersion

Creates a new version of an existing traffic policy. When you create a new version of a traffic policy, you specify the ID of the

CreateVPCAssociationAuthorization

Authorizes the AWS account that created a specified VPC to submit an AssociateVPCWithHostedZone

DeactivateKeySigningKey

Deactivates a key-signing key (KSK) so that it will not be used for signing by DNSSEC. This operation changes the

DeleteHealthCheck

Deletes a health check.

DeleteHostedZone

Deletes a hosted zone.

DeleteKeySigningKey

Deletes a key-signing key (KSK). Before you can delete a KSK, you must deactivate it. The KSK must be

DeleteQueryLoggingConfig

Deletes a configuration for DNS query logging. If you delete a configuration, Amazon Route 53 stops sending query logs to CloudWatch Logs.

DeleteReusableDelegationSet

Deletes a reusable delegation set.

DeleteTrafficPolicy

Deletes a traffic policy.

DeleteTrafficPolicyInstance

Deletes a traffic policy instance and all of the resource record sets that Amazon Route 53 created when you created the instance.

DeleteVPCAssociationAuthorization

Removes authorization to submit an AssociateVPCWithHostedZone request to associate a specified VPC

DisableHostedZoneDNSSEC

Disables DNSSEC signing in a specific hosted zone. This action does not deactivate any key-signing keys (KSKs)

DisassociateVPCFromHostedZone

Disassociates an Amazon Virtual Private Cloud (Amazon VPC) from an Amazon Route 53 private hosted zone. Note the following:

EnableHostedZoneDNSSEC

Enables DNSSEC signing in a specific hosted zone.

GetAccountLimit

Gets the specified limit for the current account, for example, the maximum number of health checks that you

GetChange

Returns the current status of a change batch request. The status is one of the

GetCheckerIpRanges

Route 53 does not perform authorization for this API because it retrieves information that is already available to the public.

GetDNSSEC

Returns information about DNSSEC for a specific hosted zone, including the key-signing keys (KSKs) in the hosted zone.

GetGeoLocation

Gets information about whether a specified geographic location is supported for Amazon Route 53 geolocation

GetHealthCheck

Gets information about a specified health check.

GetHealthCheckCount

Retrieves the number of health checks that are associated with the current AWS account.

GetHealthCheckLastFailureReason

Gets the reason that a specified health check failed most recently.

GetHealthCheckStatus

Gets status of a specified health check.

GetHostedZone

Gets information about a specified hosted zone including the four name servers assigned to the hosted zone.

GetHostedZoneCount

Retrieves the number of hosted zones that are associated with the current AWS account.

GetHostedZoneLimit

Gets the specified limit for a specified hosted zone, for example, the maximum number of records that you

GetQueryLoggingConfig

Gets information about a specified configuration for DNS query logging.

GetReusableDelegationSet

Retrieves information about a specified reusable delegation set, including the four name servers that are assigned

GetReusableDelegationSetLimit

Gets the maximum number of hosted zones that you can associate with the specified reusable delegation set.

GetTrafficPolicy

Gets information about a specific traffic policy version.

GetTrafficPolicyInstance

Gets information about a specified traffic policy instance.

GetTrafficPolicyInstanceCount

Gets the number of traffic policy instances that are associated with the current AWS account.

ListGeoLocations

Retrieves a list of supported geographic locations.

ListHealthChecks

Retrieve a list of the health checks that are associated with the current AWS account.

ListHostedZones

Retrieves a list of the public and private hosted zones that are associated with the current AWS account. The response

ListHostedZonesByName

Retrieves a list of your hosted zones in lexicographic order. The response includes a HostedZones child element

ListHostedZonesByVPC

Lists all the private hosted zones that a specified VPC is associated with, regardless of which AWS account or AWS service owns the

ListQueryLoggingConfigs

Lists the configurations for DNS query logging that are associated with the current AWS account or the configuration

ListResourceRecordSets

Lists the resource record sets in a specified hosted zone.

ListReusableDelegationSets

Retrieves a list of the reusable delegation sets that are associated with the current AWS account.

ListTagsForResource

Lists tags for one health check or hosted zone.

ListTagsForResources

Lists tags for up to 10 health checks or hosted zones.

ListTrafficPolicies

Gets information about the latest version for every traffic policy that is associated with the current AWS account.

ListTrafficPolicyInstances

Gets information about the traffic policy instances that you created by using the current AWS account.

ListTrafficPolicyInstancesByHostedZone

Gets information about the traffic policy instances that you created in a specified hosted zone.

ListTrafficPolicyInstancesByPolicy

Gets information about the traffic policy instances that you created by using a specify traffic policy version.

ListTrafficPolicyVersions

Gets information about all of the versions for a specified traffic policy.

ListVPCAssociationAuthorizations

Gets a list of the VPCs that were created by other accounts and that can be associated with a

TestDNSAnswer

Gets the value that Amazon Route 53 returns in response to a DNS request for a specified record name and type. You can optionally specify

UpdateHealthCheck

Updates an existing health check. Note that some values can't be updated.

UpdateHostedZoneComment

Updates the comment for a specified hosted zone.

UpdateTrafficPolicyComment

Updates the comment for a specified traffic policy version.

UpdateTrafficPolicyInstance

Updates the resource record sets in a specified hosted zone that were created based on the settings in a specified traffic policy version.