Module aws_sdk_route53::output

Modules

activate_key_signing_key_output

See ActivateKeySigningKeyOutput

associate_vpc_with_hosted_zone_output

See AssociateVpcWithHostedZoneOutput

change_resource_record_sets_output

See ChangeResourceRecordSetsOutput

change_tags_for_resource_output

See ChangeTagsForResourceOutput

create_health_check_output

See CreateHealthCheckOutput

create_hosted_zone_output

See CreateHostedZoneOutput

create_key_signing_key_output

See CreateKeySigningKeyOutput

create_query_logging_config_output

See CreateQueryLoggingConfigOutput

create_reusable_delegation_set_output

See CreateReusableDelegationSetOutput

create_traffic_policy_instance_output

See CreateTrafficPolicyInstanceOutput

create_traffic_policy_output

See CreateTrafficPolicyOutput

create_traffic_policy_version_output

See CreateTrafficPolicyVersionOutput

create_vpc_association_authorization_output

See CreateVpcAssociationAuthorizationOutput

deactivate_key_signing_key_output

See DeactivateKeySigningKeyOutput

delete_health_check_output

See DeleteHealthCheckOutput

delete_hosted_zone_output

See DeleteHostedZoneOutput

delete_key_signing_key_output

See DeleteKeySigningKeyOutput

delete_query_logging_config_output

See DeleteQueryLoggingConfigOutput

delete_reusable_delegation_set_output

See DeleteReusableDelegationSetOutput

delete_traffic_policy_instance_output

See DeleteTrafficPolicyInstanceOutput

delete_traffic_policy_output

See DeleteTrafficPolicyOutput

delete_vpc_association_authorization_output

See DeleteVpcAssociationAuthorizationOutput

disable_hosted_zone_dnssec_output

See DisableHostedZoneDnssecOutput

disassociate_vpc_from_hosted_zone_output

See DisassociateVpcFromHostedZoneOutput

enable_hosted_zone_dnssec_output

See EnableHostedZoneDnssecOutput

get_account_limit_output

See GetAccountLimitOutput

get_change_output

See GetChangeOutput

get_checker_ip_ranges_output

See GetCheckerIpRangesOutput

get_dnssec_output

See GetDnssecOutput

get_geo_location_output

See GetGeoLocationOutput

get_health_check_count_output

See GetHealthCheckCountOutput

get_health_check_last_failure_reason_output

See GetHealthCheckLastFailureReasonOutput

get_health_check_output

See GetHealthCheckOutput

get_health_check_status_output

See GetHealthCheckStatusOutput

get_hosted_zone_count_output

See GetHostedZoneCountOutput

get_hosted_zone_limit_output

See GetHostedZoneLimitOutput

get_hosted_zone_output

See GetHostedZoneOutput

get_query_logging_config_output

See GetQueryLoggingConfigOutput

get_reusable_delegation_set_limit_output

See GetReusableDelegationSetLimitOutput

get_reusable_delegation_set_output

See GetReusableDelegationSetOutput

get_traffic_policy_instance_count_output

See GetTrafficPolicyInstanceCountOutput

get_traffic_policy_instance_output

See GetTrafficPolicyInstanceOutput

get_traffic_policy_output

See GetTrafficPolicyOutput

list_geo_locations_output

See ListGeoLocationsOutput

list_health_checks_output

See ListHealthChecksOutput

list_hosted_zones_by_name_output

See ListHostedZonesByNameOutput

list_hosted_zones_by_vpc_output

See ListHostedZonesByVpcOutput

list_hosted_zones_output

See ListHostedZonesOutput

list_query_logging_configs_output

See ListQueryLoggingConfigsOutput

list_resource_record_sets_output

See ListResourceRecordSetsOutput

list_reusable_delegation_sets_output

See ListReusableDelegationSetsOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

list_tags_for_resources_output

See ListTagsForResourcesOutput

list_traffic_policies_output

See ListTrafficPoliciesOutput

list_traffic_policy_instances_by_hosted_zone_output

See ListTrafficPolicyInstancesByHostedZoneOutput

list_traffic_policy_instances_by_policy_output

See ListTrafficPolicyInstancesByPolicyOutput

list_traffic_policy_instances_output

See ListTrafficPolicyInstancesOutput

list_traffic_policy_versions_output

See ListTrafficPolicyVersionsOutput

list_vpc_association_authorizations_output

See ListVpcAssociationAuthorizationsOutput

test_dns_answer_output

See TestDnsAnswerOutput

update_health_check_output

See UpdateHealthCheckOutput

update_hosted_zone_comment_output

See UpdateHostedZoneCommentOutput

update_traffic_policy_comment_output

See UpdateTrafficPolicyCommentOutput

update_traffic_policy_instance_output

See UpdateTrafficPolicyInstanceOutput

Structs

ActivateKeySigningKeyOutput

AssociateVpcWithHostedZoneOutput

A complex type that contains the response information for the AssociateVPCWithHostedZone request.

ChangeResourceRecordSetsOutput

A complex type containing the response for the request.

ChangeTagsForResourceOutput

Empty response for the request.

CreateHealthCheckOutput

A complex type containing the response information for the new health check.

CreateHostedZoneOutput

A complex type containing the response information for the hosted zone.

CreateKeySigningKeyOutput

CreateQueryLoggingConfigOutput

CreateReusableDelegationSetOutput

CreateTrafficPolicyInstanceOutput

A complex type that contains the response information for the CreateTrafficPolicyInstance request.

CreateTrafficPolicyOutput

A complex type that contains the response information for the CreateTrafficPolicy request.

CreateTrafficPolicyVersionOutput

A complex type that contains the response information for the CreateTrafficPolicyVersion request.

CreateVpcAssociationAuthorizationOutput

A complex type that contains the response information from a CreateVPCAssociationAuthorization request.

DeactivateKeySigningKeyOutput

DeleteHealthCheckOutput

An empty element.

DeleteHostedZoneOutput

A complex type that contains the response to a DeleteHostedZone request.

DeleteKeySigningKeyOutput

DeleteQueryLoggingConfigOutput

DeleteReusableDelegationSetOutput

An empty element.

DeleteTrafficPolicyInstanceOutput

An empty element.

DeleteTrafficPolicyOutput

An empty element.

DeleteVpcAssociationAuthorizationOutput

Empty response for the request.

DisableHostedZoneDnssecOutput

DisassociateVpcFromHostedZoneOutput

A complex type that contains the response information for the disassociate request.

EnableHostedZoneDnssecOutput

GetAccountLimitOutput

A complex type that contains the requested limit.

GetChangeOutput

A complex type that contains the ChangeInfo element.

GetCheckerIpRangesOutput

A complex type that contains the CheckerIpRanges element.

GetDnssecOutput

GetGeoLocationOutput

A complex type that contains the response information for the specified geolocation code.

GetHealthCheckCountOutput

A complex type that contains the response to a GetHealthCheckCount request.

GetHealthCheckLastFailureReasonOutput

A complex type that contains the response to a GetHealthCheckLastFailureReason request.

GetHealthCheckOutput

A complex type that contains the response to a GetHealthCheck request.

GetHealthCheckStatusOutput

A complex type that contains the response to a GetHealthCheck

GetHostedZoneCountOutput

A complex type that contains the response to a GetHostedZoneCount request.

GetHostedZoneLimitOutput

A complex type that contains the requested limit.

GetHostedZoneOutput

A complex type that contain the response to a GetHostedZone request.

GetQueryLoggingConfigOutput

GetReusableDelegationSetLimitOutput

A complex type that contains the requested limit.

GetReusableDelegationSetOutput

A complex type that contains the response to the GetReusableDelegationSet request.

GetTrafficPolicyInstanceCountOutput

A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

GetTrafficPolicyInstanceOutput

A complex type that contains information about the resource record sets that Amazon Route 53 created based on a specified traffic policy.

GetTrafficPolicyOutput

A complex type that contains the response information for the request.

ListGeoLocationsOutput

A complex type containing the response information for the request.

ListHealthChecksOutput

A complex type that contains the response to a ListHealthChecks request.

ListHostedZonesByNameOutput

A complex type that contains the response information for the request.

ListHostedZonesByVpcOutput

ListHostedZonesOutput

ListQueryLoggingConfigsOutput

ListResourceRecordSetsOutput

A complex type that contains list information for the resource record set.

ListReusableDelegationSetsOutput

A complex type that contains information about the reusable delegation sets that are associated with the current AWS account.

ListTagsForResourceOutput

A complex type that contains information about the health checks or hosted zones for which you want to list tags.

ListTagsForResourcesOutput

A complex type containing tags for the specified resources.

ListTrafficPoliciesOutput

A complex type that contains the response information for the request.

ListTrafficPolicyInstancesByHostedZoneOutput

A complex type that contains the response information for the request.

ListTrafficPolicyInstancesByPolicyOutput

A complex type that contains the response information for the request.

ListTrafficPolicyInstancesOutput

A complex type that contains the response information for the request.

ListTrafficPolicyVersionsOutput

A complex type that contains the response information for the request.

ListVpcAssociationAuthorizationsOutput

A complex type that contains the response information for the request.

TestDnsAnswerOutput

A complex type that contains the response to a TestDNSAnswer request.

UpdateHealthCheckOutput

A complex type that contains the response to the UpdateHealthCheck request.

UpdateHostedZoneCommentOutput

A complex type that contains the response to the UpdateHostedZoneComment request.

UpdateTrafficPolicyCommentOutput

A complex type that contains the response information for the traffic policy.

UpdateTrafficPolicyInstanceOutput

A complex type that contains information about the resource record sets that Amazon Route 53